:recycle: Rename to Discard Dialog
